+++ transport/register.go
@@ -21,6 +21,7 @@
 package transport
 
 import (
+	"fmt"
 	"sort"
 
 	"go.uber.org/yarpc/internal/errors"
@@ -46,8 +47,8 @@ type Registrant struct {
 	// Name of the procedure.
 	Procedure string
 
-	// Handler implementing the given procedure.
-	Handler UnaryHandler
+	// HandlerSpec specifiying which handler and rpc type.
+	HandlerSpec HandlerSpec
 }
 
 // Registry maintains and provides access to a collection of procedures and
@@ -66,14 +67,15 @@ type Registry interface {
 	//
 	// service may be empty to indicate that the default service name should
 	// be used.
-	GetHandler(service, procedure string) (UnaryHandler, error)
+	GetHandlerSpec(service, procedure string) (HandlerSpec, error)
 }
 
 // MapRegistry is a Registry that maintains a map of the registered
 // procedures.
 type MapRegistry struct {
 	defaultService string
-	unaryEntries   map[ServiceProcedure]UnaryHandler
+	entries        map[ServiceProcedure]UnaryHandler
+	onewayEntries  map[ServiceProcedure]OnewayHandler
 }
 
 // NewMapRegistry builds a new MapRegistry that uses the given name as the
@@ -81,7 +83,8 @@ type MapRegistry struct {
 func NewMapRegistry(defaultService string) MapRegistry {
 	return MapRegistry{
 		defaultService: defaultService,
-		unaryEntries:   make(map[ServiceProcedure]UnaryHandler),
+		entries:        make(map[ServiceProcedure]UnaryHandler),
+		onewayEntries:  make(map[ServiceProcedure]OnewayHandler),
 	}
 }
 
@@ -92,33 +95,43 @@ func (m MapRegistry) Register(rs []Registrant) {
 			r.Service = m.defaultService
 		}
 
-		m.unaryEntries[ServiceProcedure{r.Service, r.Procedure}] = r.Handler
+		switch r.HandlerSpec.Type {
+		case Unary:
+			m.entries[ServiceProcedure{r.Service, r.Procedure}] = r.HandlerSpec.UnaryHandler
+		case Oneway:
+			m.onewayEntries[ServiceProcedure{r.Service, r.Procedure}] = r.HandlerSpec.OnewayHandler
+		default:
+			panic(fmt.Sprintf("Unknown RPC Type %v, for %s::%s", r.HandlerSpec.Type, r.Service, r.Procedure))
+		}
 	}
 }
 
 // ServiceProcedures returns a list of services and their procedures that
 // have been registered so far.
 func (m MapRegistry) ServiceProcedures() []ServiceProcedure {
-	procs := make([]ServiceProcedure, 0, len(m.unaryEntries))
-	for k := range m.unaryEntries {
+	procs := make([]ServiceProcedure, 0, len(m.entries))
+	for k := range m.entries {
 		procs = append(procs, k)
 	}
 	sort.Sort(byServiceProcedure(procs))
 	return procs
 }
 
-// GetHandler retrieves the Handler for the given Procedure or returns an
+// GetHandlerSpec retrieves the HandlerSpec for the given Procedure or returns an
 // error.
-func (m MapRegistry) GetHandler(service, procedure string) (UnaryHandler, error) {
+func (m MapRegistry) GetHandlerSpec(service, procedure string) (HandlerSpec, error) {
 	if service == "" {
 		service = m.defaultService
 	}
 
-	if h, ok := m.unaryEntries[ServiceProcedure{service, procedure}]; ok {
-		return h, nil
+	if h, ok := m.entries[ServiceProcedure{service, procedure}]; ok {
+		return HandlerSpec{Type: Unary, UnaryHandler: h}, nil
+	}
+	if h, ok := m.onewayEntries[ServiceProcedure{service, procedure}]; ok {
+		return HandlerSpec{Type: Oneway, OnewayHandler: h}, nil
 	}
 
-	return nil, errors.UnrecognizedProcedureError{
+	return HandlerSpec{}, errors.UnrecognizedProcedureError{
 		Service:   service,
 		Procedure: procedure,
 	}